refactor: Phase 5 [5.2] migrate models/ — move mtl_poi.py, add deprecation shims (BPR)
